+++ src/parser.js
@@ -1,4 +1,4 @@
-/*eslint-disable*/
+/* eslint-disable strict, no-eval*/
 
 /*
 	This may well be the worst code I've ever written.
@@ -8,23 +8,24 @@
 	It's also the reason you should never use panic
 	in user-facing code.
 
-	The purpose of the parser is to take a stringified
-	function, a scope, and return a function that has
-	locally scoped variables according to the properties
-	in the object (second param).
+	This module takes a function string and magically
+	(dangerously) turns it into a function using `eval`.
+	The variables are named all shouty because eval can see
+	local scope variables.
 */
 
 /**
  * Take a function string and evaluate it,
  * optionally injecting local variables.
- * Includes hackery for IE6 compatibility.
- * @return {Function} - The evaluated function.
+ * @param {String} PANIC_SRC_STR - A function string to parse.
+ * @return {Function} - The resulting function reference.
  */
-module.exports = function () {
-	var arguments = module.exports.arguments || arguments;
+module.exports = function (PANIC_SRC_STR) {
+
+	/** IE6 doesn't like anon function expressions. */
 	var PANIC_CB_FUNCTION;
-	with (arguments[1] || {}) {
-		eval('PANIC_CB_FUNCTION = ' + arguments[0]);
-		return PANIC_CB_FUNCTION;
-	}
+
+	eval('PANIC_CB_FUNCTION = ' + PANIC_SRC_STR);
+
+	return PANIC_CB_FUNCTION;
 };