+++ test/parser.js
@@ -3,33 +3,29 @@
 'use strict';
 
 var parser = require('../src/parser');
-var expect = require('chai').expect;
+var expect = require('expect');
 
 describe('The parser', function () {
-	var cb = (function () {
-		return 'success';
-	}).toString();
 
-	it('should return an evaluated string', function () {
-		expect(parser('5')).to.eq(5);
+	it('should parse values', function () {
+		var value = parser('5');
+		expect(value).toBe(5);
 	});
 
-	it('should return a function when given a function string', function () {
-		expect(parser(cb)()).to.eq('success');
+	it('should parse expressions', function () {
+		var value = parser('10 + 5');
+		expect(value).toBe(15);
 	});
 
-	it('should allow dynamic scope variables', function () {
-		var mutable;
-		parser(function () {
-			mutable.called = true;
-			expect(scopedVariable).to.eq('yep');
-		}.toString(), {
-			scopedVariable: 'yep',
-			expect: expect,
-			mutable: mutable = {
-				called: false
-			}
-		})();
-		expect(mutable.called).to.eq(true);
+	it('should parse function strings', function () {
+		var str = String(function () {
+			return 'success!';
+		});
+
+		var fn = parser(str);
+		var result = fn();
+
+		expect(result).toBe('success!');
 	});
+
 });